+++ osmoutils/accum/accum_test.go
@@ -4,15 +4,15 @@ import (
 	"testing"
 
 	"github.com/cosmos/cosmos-sdk/store"
+	iavlstore "github.com/cosmos/cosmos-sdk/store/iavl"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/iavl"
+	"github.com/gogo/protobuf/proto"
 	"github.com/stretchr/testify/suite"
 	dbm "github.com/tendermint/tm-db"
 
 	"github.com/osmosis-labs/osmosis/osmoutils"
 	accumPackage "github.com/osmosis-labs/osmosis/osmoutils/accum"
-
-	iavlstore "github.com/cosmos/cosmos-sdk/store/iavl"
 )
 
 type AccumTestSuite struct {
@@ -22,11 +22,11 @@ type AccumTestSuite struct {
 }
 
 var (
-	testAddressOne   = sdk.AccAddress([]byte("addr1_______________"))
-	testAddressTwo   = sdk.AccAddress([]byte("addr2_______________"))
-	testAddressThree = sdk.AccAddress([]byte("addr3_______________"))
+	testAddressOne   = sdk.AccAddress([]byte("addr1_______________")).String()
+	testAddressTwo   = sdk.AccAddress([]byte("addr2_______________")).String()
+	testAddressThree = sdk.AccAddress([]byte("addr3_______________")).String()
 
-	emptyPositionOptions = accumPackage.PositionOptions{}
+	emptyPositionOptions = accumPackage.Options{}
 	testNameOne          = "myaccumone"
 	testNameTwo          = "myaccumtwo"
 	testNameThree        = "myaccumthree"
@@ -151,44 +151,40 @@ func (suite *AccumTestSuite) TestNewPosition() {
 
 	tests := map[string]struct {
 		accObject        accumPackage.AccumulatorObject
-		addr             sdk.AccAddress
+		name             string
 		numShareUnits    sdk.Dec
-		options          accumPackage.PositionOptions
+		options          *accumPackage.Options
 		expectedPosition accumPackage.Record
 	}{
 		"test address one - position created": {
 			accObject:        accObject,
-			addr:             testAddressOne,
+			name:             testAddressOne,
 			numShareUnits:    positionOne.NumShares,
-			options:          emptyPositionOptions,
 			expectedPosition: positionOne,
 		},
-		"test address two - position created": {
+		"test address two (non-nil options) - position created": {
 			accObject:        accObject,
-			addr:             testAddressTwo,
+			name:             testAddressTwo,
 			numShareUnits:    positionTwo.NumShares,
-			options:          emptyPositionOptions,
 			expectedPosition: positionTwo,
+			options:          &emptyPositionOptions,
 		},
 		"test address one - position overwritten": {
 			accObject:        accObject,
-			addr:             testAddressOne,
+			name:             testAddressOne,
 			numShareUnits:    positionOneV2.NumShares,
-			options:          emptyPositionOptions,
 			expectedPosition: positionOneV2,
 		},
 		"test address three - added": {
 			accObject:        accObject,
-			addr:             testAddressThree,
+			name:             testAddressThree,
 			numShareUnits:    positionThree.NumShares,
-			options:          emptyPositionOptions,
 			expectedPosition: positionThree,
 		},
 		"test address one with non-empty accumulator - position created": {
 			accObject:        accumPackage.CreateRawAccumObject(suite.store, testNameTwo, initialCoinsDenomOne),
-			addr:             testAddressOne,
+			name:             testAddressOne,
 			numShareUnits:    positionOne.NumShares,
-			options:          emptyPositionOptions,
 			expectedPosition: withInitialAccumValue(positionOne, initialCoinsDenomOne),
 		},
 	}
@@ -197,10 +193,10 @@ func (suite *AccumTestSuite) TestNewPosition() {
 		tc := tc
 		suite.Run(name, func() {
 			// System under test.
-			tc.accObject.NewPosition(tc.addr, tc.numShareUnits, tc.options)
+			tc.accObject.NewPosition(tc.name, tc.numShareUnits, tc.options)
 
 			// Assertions.
-			positions := tc.accObject.GetPosition(tc.addr)
+			positions := tc.accObject.GetPosition(tc.name)
 			suite.Require().Equal(tc.expectedPosition, positions)
 		})
 	}
@@ -215,6 +211,12 @@ func (suite *AccumTestSuite) TestClaimRewards() {
 			sdk.NewDecCoinFromDec(denomTwo, sdk.OneDec())).MulDec(sdk.NewDec(3))
 	)
 
+	// single output convenience wrapper.
+	toCoins := func(decCoins sdk.DecCoins) sdk.Coins {
+		coins, _ := decCoins.TruncateDecimal()
+		return coins
+	}
+
 	// We setup store and accum
 	// once at beginning so we can test duplicate positions
 	suite.SetupTest()
@@ -225,8 +227,8 @@ func (suite *AccumTestSuite) TestClaimRewards() {
 	accumNoRewards := accumPackage.CreateRawAccumObject(suite.store, testNameOne, emptyCoins)
 
 	// Create positions at testAddressOne and testAddressTwo.
-	accumNoRewards.NewPosition(testAddressOne, positionOne.NumShares, emptyPositionOptions)
-	accumNoRewards.NewPosition(testAddressTwo, positionTwo.NumShares, emptyPositionOptions)
+	accumNoRewards.NewPosition(testAddressOne, positionOne.NumShares, nil)
+	accumNoRewards.NewPosition(testAddressTwo, positionTwo.NumShares, nil)
 
 	// 2. One accumulator reward coin, 1 position accumulator, no unclaimed rewards in position.
 	accumOneReward := accumPackage.CreateRawAccumObject(suite.store, testNameTwo, initialCoinsDenomOne)
@@ -245,54 +247,51 @@ func (suite *AccumTestSuite) TestClaimRewards() {
 	accumThreeRewards = accumPackage.WithPosition(accumThreeRewards, testAddressOne, withUnclaimedRewards(positionOne, initialCoinsDenomOne))
 
 	// Create positions at testAddressThree with no unclaimed rewards.
-	accumThreeRewards.NewPosition(testAddressTwo, positionTwo.NumShares, emptyPositionOptions)
+	accumThreeRewards.NewPosition(testAddressTwo, positionTwo.NumShares, nil)
 
 	// Triple the accumulator value.
 	accumThreeRewards.SetValue(tripleDenomOneAndTwo)
 
 	tests := map[string]struct {
 		accObject      accumPackage.AccumulatorObject
-		addr           sdk.AccAddress
-		expectedResult sdk.DecCoins
+		name           string
+		expectedResult sdk.Coins
 		expectError    error
 	}{
 		"claim at testAddressOne with no rewards - success": {
 			accObject:      accumNoRewards,
-			addr:           testAddressOne,
-			expectedResult: emptyCoins,
+			name:           testAddressOne,
+			expectedResult: toCoins(emptyCoins),
 		},
 		"claim at testAddressTwo with no rewards - success": {
 			accObject:      accumNoRewards,
-			addr:           testAddressTwo,
-			expectedResult: emptyCoins,
+			name:           testAddressTwo,
+			expectedResult: toCoins(emptyCoins),
 		},
 		"claim at testAddressTwo with no rewards - error - no position": {
 			accObject:   accumNoRewards,
-			addr:        testAddressThree,
-			expectError: accumPackage.NoPositionError{Address: testAddressThree},
+			name:        testAddressThree,
+			expectError: accumPackage.NoPositionError{Name: testAddressThree},
 		},
 		"claim at testAddressThree with single reward token - success": {
 			accObject: accumOneReward,
-			addr:      testAddressThree,
+			name:      testAddressThree,
 			// denomOne: (200.2 - 100.1) * 300 (accum diff * share count) = 30030
-			expectedResult: initialCoinsDenomOne.MulDec(positionThree.NumShares),
+			expectedResult: toCoins(initialCoinsDenomOne.MulDec(positionThree.NumShares)),
 		},
 		"claim at testAddressOne with multiple reward tokens and unclaimed rewards - success": {
 			accObject: accumThreeRewards,
-			addr:      testAddressOne,
+			name:      testAddressOne,
 			// denomOne: (300.3 - 0) * 100 (accum diff * share count) + 100.1 (unclaimed rewards) = 30130.1
 			// denomTwo: (3 - 0) * 100 (accum diff * share count) = 300
-			expectedResult: tripleDenomOneAndTwo.MulDec(positionOne.NumShares).Add(initialCoinDenomOne),
+			expectedResult: toCoins(tripleDenomOneAndTwo.MulDec(positionOne.NumShares).Add(initialCoinDenomOne)),
 		},
 		"claim at testAddressTwo with multiple reward tokens and no unclaimed rewards - success": {
 			accObject: accumThreeRewards,
-			addr:      testAddressTwo,
+			name:      testAddressTwo,
 			// denomOne: (300.3 - 0) * 200 (accum diff * share count) = 60060.6
 			// denomTwo: (3 - 0) * 200  (accum diff * share count) = 600
-			expectedResult: sdk.NewDecCoins(
-				initialCoinDenomOne,
-				sdk.NewDecCoinFromDec(denomTwo, sdk.OneDec()),
-			).MulDec(positionTwo.NumShares).MulDec(sdk.NewDec(3)),
+			expectedResult: toCoins(sdk.NewDecCoins(initialCoinDenomOne, sdk.NewDecCoinFromDec(denomTwo, sdk.OneDec())).MulDec(positionTwo.NumShares).MulDec(sdk.NewDec(3))),
 		},
 	}
 
@@ -300,7 +299,7 @@ func (suite *AccumTestSuite) TestClaimRewards() {
 		tc := tc
 		suite.Run(name, func() {
 			// System under test.
-			actualResult, err := tc.accObject.ClaimRewards(tc.addr)
+			actualResult, err := tc.accObject.ClaimRewards(tc.name)
 
 			// Assertions.
 
@@ -314,11 +313,11 @@ func (suite *AccumTestSuite) TestClaimRewards() {
 
 			suite.Require().Equal(tc.expectedResult, actualResult)
 
-			finalPosition := tc.accObject.GetPosition(tc.addr)
+			finalPosition := tc.accObject.GetPosition(tc.name)
 			suite.Require().NoError(err)
 
 			// Unclaimed rewards are reset.
-			suite.Require().True(finalPosition.UnclaimedRewards.IsZero())
+			suite.Require().Equal(emptyCoins, finalPosition.UnclaimedRewards)
 		})
 	}
 }
@@ -457,27 +456,27 @@ func (suite *AccumTestSuite) TestAddToPosition() {
 		suite.Run(name, func() {
 			// We reset the store for each test
 			suite.SetupTest()
-			addr := osmoutils.CreateRandomAccounts(1)[0]
+			positionName := osmoutils.CreateRandomAccounts(1)[0].String()
 
 			// Create a new accumulator with initial value specified by test case
 			curAccum := accumPackage.CreateRawAccumObject(suite.store, testNameOne, tc.accumInit)
 
 			// Create new position in store (raw to minimize dependencies)
 			if !tc.addrDoesNotExist {
-				accumPackage.CreateRawPosition(curAccum, addr, tc.startingNumShares, tc.startingUnclaimedRewards, emptyPositionOptions)
+				accumPackage.CreateRawPosition(curAccum, positionName, tc.startingNumShares, tc.startingUnclaimedRewards, nil)
 			}
 
 			// Update accumulator with expAccumDelta (increasing position's rewards by a proportional amount)
 			curAccum = accumPackage.CreateRawAccumObject(suite.store, testNameOne, tc.accumInit.Add(tc.expAccumDelta...))
 
 			// Add newShares to position
-			err := curAccum.AddToPosition(addr, tc.newShares)
+			err := curAccum.AddToPosition(positionName, tc.newShares)
 
 			if tc.expPass {
 				suite.Require().NoError(err)
 
 				// Get updated position for comparison
-				newPosition, err := accumPackage.GetPosition(curAccum, addr)
+				newPosition, err := accumPackage.GetPosition(curAccum, positionName)
 				suite.Require().NoError(err)
 
 				// Ensure position's accumulator value is moved up to init + delta
@@ -499,7 +498,7 @@ func (suite *AccumTestSuite) TestAddToPosition() {
 				// Further checks to ensure state was not mutated upon error
 				if !tc.addrDoesNotExist {
 					// Get new position for comparison
-					newPosition, err := accumPackage.GetPosition(curAccum, addr)
+					newPosition, err := accumPackage.GetPosition(curAccum, positionName)
 					suite.Require().NoError(err)
 
 					// Ensure that numShares, accumulator value, and unclaimed rewards are unchanged
@@ -645,27 +644,27 @@ func (suite *AccumTestSuite) TestRemoveFromPosition() {
 		suite.Run(name, func() {
 			// We reset the store for each test
 			suite.SetupTest()
-			addr := osmoutils.CreateRandomAccounts(1)[0]
+			positionName := osmoutils.CreateRandomAccounts(1)[0].String()
 
 			// Create a new accumulator with initial value specified by test case
 			curAccum := accumPackage.CreateRawAccumObject(suite.store, testNameOne, tc.accumInit)
 
 			// Create new position in store (raw to minimize dependencies)
 			if !tc.addrDoesNotExist {
-				accumPackage.CreateRawPosition(curAccum, addr, tc.startingNumShares, tc.startingUnclaimedRewards, emptyPositionOptions)
+				accumPackage.CreateRawPosition(curAccum, positionName, tc.startingNumShares, tc.startingUnclaimedRewards, nil)
 			}
 
 			// Update accumulator with expAccumDelta (increasing position's rewards by a proportional amount)
 			curAccum = accumPackage.CreateRawAccumObject(suite.store, testNameOne, tc.accumInit.Add(tc.expAccumDelta...))
 
 			// Remove removedShares from position
-			err := curAccum.RemoveFromPosition(addr, tc.removedShares)
+			err := curAccum.RemoveFromPosition(positionName, tc.removedShares)
 
 			if tc.expPass {
 				suite.Require().NoError(err)
 
 				// Get updated position for comparison
-				newPosition, err := accumPackage.GetPosition(curAccum, addr)
+				newPosition, err := accumPackage.GetPosition(curAccum, positionName)
 				suite.Require().NoError(err)
 
 				// Ensure position's accumulator value is moved up to init + delta
@@ -691,7 +690,7 @@ func (suite *AccumTestSuite) TestRemoveFromPosition() {
 				// Further checks to ensure state was not mutated upon error
 				if !tc.addrDoesNotExist {
 					// Get new position for comparison
-					newPosition, err := accumPackage.GetPosition(curAccum, addr)
+					newPosition, err := accumPackage.GetPosition(curAccum, positionName)
 					suite.Require().NoError(err)
 
 					// Ensure that numShares, accumulator value, and unclaimed rewards are unchanged
@@ -756,27 +755,27 @@ func (suite *AccumTestSuite) TestGetPositionSize() {
 		suite.Run(name, func() {
 			// We reset the store for each test
 			suite.SetupTest()
-			addr := osmoutils.CreateRandomAccounts(1)[0]
+			positionName := osmoutils.CreateRandomAccounts(1)[0].String()
 
 			// Create a new accumulator with initial value specified by test case
 			curAccum := accumPackage.CreateRawAccumObject(suite.store, testNameOne, tc.accumInit)
 
 			// Create new position in store (raw to minimize dependencies)
 			if !tc.addrDoesNotExist {
-				accumPackage.CreateRawPosition(curAccum, addr, tc.numShares, sdk.NewDecCoins(), emptyPositionOptions)
+				accumPackage.CreateRawPosition(curAccum, positionName, tc.numShares, sdk.NewDecCoins(), nil)
 			}
 
 			// Update accumulator with expAccumDelta (increasing position's rewards by a proportional amount)
 			curAccum = accumPackage.CreateRawAccumObject(suite.store, testNameOne, tc.accumInit.Add(tc.expAccumDelta...))
 
 			// Get position size from valid address (or from nonexistant if address does not exist)
-			positionSize, err := curAccum.GetPositionSize(addr)
+			positionSize, err := curAccum.GetPositionSize(positionName)
 
 			if tc.changedShares.GT(sdk.ZeroDec()) {
-				accumPackage.CreateRawPosition(curAccum, addr, tc.numShares.Add(tc.changedShares), sdk.NewDecCoins(), emptyPositionOptions)
+				accumPackage.CreateRawPosition(curAccum, positionName, tc.numShares.Add(tc.changedShares), sdk.NewDecCoins(), nil)
 			}
 
-			positionSize, err = curAccum.GetPositionSize(addr)
+			positionSize, err = curAccum.GetPositionSize(positionName)
 
 			if tc.expPass {
 				suite.Require().NoError(err)
@@ -789,6 +788,81 @@ func (suite *AccumTestSuite) TestGetPositionSize() {
 			} else {
 				suite.Require().Error(err)
 			}
+		})
+	}
+}
+
+// TestMarhsalUnmarshalRecord displays that we may use Records without options
+// For records with nil options, adding new fields to `Options`, should not
+// require future migrations.
+func (suite *AccumTestSuite) TestMarhsalUnmarshalRecord() {
+	suite.SetupTest()
+
+	recordNoOptions := accumPackage.Record{
+		NumShares: sdk.OneDec(),
+		InitAccumValue: sdk.NewDecCoins(
+			sdk.NewDecCoinFromDec(denomOne, sdk.OneDec()),
+		),
+		UnclaimedRewards: sdk.NewDecCoins(
+			sdk.NewDecCoinFromDec(denomOne, sdk.OneDec()),
+		),
+	}
+
+	bz, err := proto.Marshal(&recordNoOptions)
+	suite.Require().NoError(err)
+
+	var unmarshaledRecord accumPackage.Record
+	proto.Unmarshal(bz, &unmarshaledRecord)
+	// Options should be nil, not an empty struct
+	suite.Require().True(unmarshaledRecord.Options == nil)
+}
+
+func (suite *AccumTestSuite) TestUpdateAccumulator() {
+	tests := map[string]struct {
+		updateAmmount sdk.DecCoins
+
+		expectedValue sdk.DecCoins
+	}{
+		"positive": {
+			updateAmmount: initialCoinsDenomOne,
+
+			expectedValue: initialCoinsDenomOne,
+		},
+		"negative": {
+			updateAmmount: initialCoinsDenomOne.MulDec(sdk.NewDec(-1)),
+
+			expectedValue: initialCoinsDenomOne.MulDec(sdk.NewDec(-1)),
+		},
+		"multiple coins": {
+			updateAmmount: initialCoinsDenomOne.Add(initialCoinDenomTwo),
+
+			expectedValue: initialCoinsDenomOne.Add(initialCoinDenomTwo),
+		},
+	}
+
+	for name, tc := range tests {
+		suite.Run(name, func() {
+			// Setup
+			suite.SetupTest()
+
+			err := accumPackage.MakeAccumulator(suite.store, testNameOne)
+			suite.Require().NoError(err)
+			originalAccum, err := accumPackage.GetAccumulator(suite.store, testNameOne)
+			suite.Require().NoError(err)
+
+			// System under test.
+			originalAccum.UpdateAccumulator(tc.updateAmmount)
+
+			// Validations.
+
+			// validate that the reciever is mutated.
+			suite.Require().Equal(tc.expectedValue, originalAccum.GetValue())
+
+			accumFromStore, err := accumPackage.GetAccumulator(suite.store, testNameOne)
+			suite.Require().NoError(err)
+
+			// validate that store is updated.
+			suite.Require().Equal(tc.expectedValue, accumFromStore.GetValue())
 		})
 	}
 }